Update pfc transducer connection
PFC send sInstantSpeed to Transducer
